Fix (Subject, SubjectTeacher): Ahora no se pueden eliminar asignaturas asociadas a profesores, ni se pueden eliminar registros asignatura-profesor con calificaciones asociadas

+++ app/Services/Subjects/DeleteSubjectService.php
@@ -10,6 +10,14 @@ class DeleteSubjectService
     public function handle(Subject $subject): void
     {
         DB::transaction(function () use ($subject) {
+            if ($subject->sectionSubjectTeachers()->exists()) {
+                throw new \Exception('No se puede eliminar una asignatura con asignaciones en secciones.');
+            }
+
+            if ($subject->teachers()->exists()) {
+                throw new \Exception('No se puede eliminar una asignatura con profesores asignados.');
+            }
+
             $subject->delete();
         });
     }
